requirements.txt
torch
torchvision
onnxruntime-gpu
numpy
Pillow
requests
tqdm
